# MQE: Add support for avg aggregation