--- a/pkg/streamingpromql/aggregations/common.go
+++ b/pkg/streamingpromql/aggregations/common.go
@@ -22,6 +22,7 @@ type AggregationGroup interface {
 type AggregationGroupFactory func() AggregationGroup
 
 var AggregationGroupFactories = map[parser.ItemType]AggregationGroupFactory{
+	parser.AVG: func() AggregationGroup { return &AvgAggregationGroup{} },
 	parser.MAX: func() AggregationGroup { return NewMinMaxAggregationGroup(true) },
 	parser.MIN: func() AggregationGroup { return NewMinMaxAggregationGroup(false) },
 	parser.SUM: func() AggregationGroup { return &SumAggregationGroup{} },