--- a/pkg/streamingpromql/aggregations/min_max.go
+++ b/pkg/streamingpromql/aggregations/min_max.go
@@ -49,7 +49,7 @@ func (g *MinMaxAggregationGroup) minAccumulatePoint(idx int64, f float64) {
 
 func (g *MinMaxAggregationGroup) AccumulateSeries(data types.InstantVectorSeriesData, steps int, start int64, interval int64, memoryConsumptionTracker *limiting.MemoryConsumptionTracker, _ functions.EmitAnnotationFunc) error {
 	if (len(data.Floats) > 0 || len(data.Histograms) > 0) && g.floatValues == nil {
-		// Even if we have histograms, we have to populate the float slices, as we'll treat histograms as if they have value 0.
+		// Even if we only have histograms, we have to populate the float slices, as we'll treat histograms as if they have value 0.
 		// This is consistent with Prometheus but may not be the desired value: https://github.com/prometheus/prometheus/issues/14711
 
 		var err error