@jhesketh 2024-09-04
Benchmark

<details>

```

                                                            │      Prometheus      │               Mimir                │
                                                            │        sec/op        │   sec/op     vs base               │
Query/avg(a_1),_instant_query                                          685.2µ ± 1%   668.1µ ± 1%   -2.50% (p=0.002 n=6)
Query/avg(a_1),_range_query_with_100_steps                             723.2µ ± 1%   688.4µ ± 1%   -4.81% (p=0.002 n=6)
Query/avg(a_1),_range_query_with_1000_steps                            933.7µ ± 1%   855.1µ ± 1%   -8.41% (p=0.002 n=6)
Query/avg(a_100),_instant_query                                        2.288m ± 1%   2.189m ± 1%   -4.33% (p=0.002 n=6)
Query/avg(a_100),_range_query_with_100_steps                           3.623m ± 1%   3.578m ± 1%   -1.23% (p=0.009 n=6)
Query/avg(a_100),_range_query_with_1000_steps                          16.43m ± 1%   16.11m ± 1%   -1.95% (p=0.002 n=6)
Query/avg(a_2000),_instant_query                                       26.56m ± 2%   25.47m ± 1%   -4.10% (p=0.002 n=6)
Query/avg(a_2000),_range_query_with_100_steps                          55.04m ± 1%   52.72m ± 1%   -4.21% (p=0.002 n=6)
Query/avg(a_2000),_range_query_with_1000_steps                         344.0m ± 1%   284.2m ± 1%  -17.39% (p=0.002 n=6)
Query/avg_without_(l)(h_1),_instant_query                              808.7µ ± 1%   802.8µ ± 1%        ~ (p=0.093 n=6)
Query/avg_without_(l)(h_1),_range_query_with_100_steps                 930.5µ ± 1%   902.2µ ± 1%   -3.04% (p=0.002 n=6)
Query/avg_without_(l)(h_1),_range_query_with_1000_steps                1.832m ± 0%   1.687m ± 0%   -7.95% (p=0.002 n=6)
Query/avg_without_(l)(h_100),_instant_query                            9.451m ± 1%   9.115m ± 2%   -3.55% (p=0.002 n=6)
Query/avg_without_(l)(h_100),_range_query_with_100_steps               18.06m ± 1%   17.42m ± 1%   -3.56% (p=0.002 n=6)
Query/avg_without_(l)(h_100),_range_query_with_1000_steps              94.31m ± 3%   90.19m ± 0%   -4.37% (p=0.002 n=6)
Query/avg_without_(l)(h_2000),_instant_query                           172.6m ± 1%   165.2m ± 1%   -4.27% (p=0.002 n=6)
Query/avg_without_(l)(h_2000),_range_query_with_100_steps              361.1m ± 2%   322.2m ± 1%  -10.77% (p=0.002 n=6)
Query/avg_without_(l)(h_2000),_range_query_with_1000_steps              2.845 ± 1%    1.717 ± 1%  -39.66% (p=0.002 n=6)
Query/avg_without_(le)(h_1),_instant_query                             811.9µ ± 1%   796.5µ ± 1%   -1.89% (p=0.002 n=6)
Query/avg_without_(le)(h_1),_range_query_with_100_steps                919.5µ ± 1%   896.7µ ± 1%   -2.47% (p=0.002 n=6)
Query/avg_without_(le)(h_1),_range_query_with_1000_steps               1.744m ± 1%   1.684m ± 0%   -3.42% (p=0.002 n=6)
Query/avg_without_(le)(h_100),_instant_query                           9.657m ± 1%   9.276m ± 1%   -3.94% (p=0.002 n=6)
Query/avg_without_(le)(h_100),_range_query_with_100_steps              18.31m ± 1%   17.61m ± 1%   -3.80% (p=0.002 n=6)
Query/avg_without_(le)(h_100),_range_query_with_1000_steps             99.98m ± 4%   89.80m ± 0%  -10.19% (p=0.002 n=6)
Query/avg_without_(le)(h_2000),_instant_query                          176.7m ± 1%   170.3m ± 1%   -3.64% (p=0.002 n=6)
Query/avg_without_(le)(h_2000),_range_query_with_100_steps             379.0m ± 1%   333.4m ± 1%  -12.02% (p=0.002 n=6)
Query/avg_without_(le)(h_2000),_range_query_with_1000_steps             3.193 ± 2%    1.745 ± 1%  -45.36% (p=0.002 n=6)
Query/avg(nh_1),_instant_query                                         830.8µ ± 1%   808.8µ ± 1%   -2.65% (p=0.002 n=6)
Query/avg(nh_1),_range_query_with_100_steps                            980.5µ ± 1%   921.8µ ± 1%   -5.98% (p=0.002 n=6)
Query/avg(nh_1),_range_query_with_1000_steps                           2.235m ± 1%   1.861m ± 4%  -16.70% (p=0.002 n=6)
Query/avg(nh_100),_instant_query                                       12.11m ± 1%   11.89m ± 0%   -1.85% (p=0.002 n=6)
Query/avg(nh_100),_range_query_with_100_steps                          30.36m ± 4%   23.50m ± 3%  -22.61% (p=0.002 n=6)
Query/avg(nh_100),_range_query_with_1000_steps                         267.6m ± 2%   130.6m ± 1%  -51.21% (p=0.002 n=6)
Query/avg(nh_2000),_instant_query                                      220.4m ± 1%   216.3m ± 1%   -1.88% (p=0.002 n=6)
Query/avg(nh_2000),_range_query_with_100_steps                         778.3m ± 2%   441.8m ± 1%  -43.23% (p=0.002 n=6)
Query/avg(nh_2000),_range_query_with_1000_steps                         6.749 ± 3%    2.596 ± 2%  -61.53% (p=0.002 n=6)
Query/avg_by_(l)(h_1),_instant_query                                   806.3µ ± 4%   800.9µ ± 2%        ~ (p=0.180 n=6)
Query/avg_by_(l)(h_1),_range_query_with_100_steps                      919.3µ ± 1%   891.5µ ± 2%   -3.03% (p=0.002 n=6)
Query/avg_by_(l)(h_1),_range_query_with_1000_steps                     1.744m ± 1%   1.682m ± 1%   -3.56% (p=0.002 n=6)
Query/avg_by_(l)(h_100),_instant_query                                 9.526m ± 1%   9.296m ± 1%   -2.41% (p=0.002 n=6)
Query/avg_by_(l)(h_100),_range_query_with_100_steps                    18.30m ± 0%   17.64m ± 1%   -3.64% (p=0.002 n=6)
Query/avg_by_(l)(h_100),_range_query_with_1000_steps                   99.45m ± 2%   89.88m ± 0%   -9.63% (p=0.002 n=6)
Query/avg_by_(l)(h_2000),_instant_query                                175.0m ± 2%   169.6m ± 1%   -3.07% (p=0.002 n=6)
Query/avg_by_(l)(h_2000),_range_query_with_100_steps                   379.8m ± 1%   332.6m ± 0%  -12.41% (p=0.002 n=6)
Query/avg_by_(l)(h_2000),_range_query_with_1000_steps                   3.200 ± 2%    1.739 ± 1%  -45.66% (p=0.002 n=6)
Query/avg_by_(le)(h_1),_instant_query                                  823.8µ ± 4%   816.4µ ± 3%   -0.90% (p=0.004 n=6)
Query/avg_by_(le)(h_1),_range_query_with_100_steps                     934.5µ ± 1%   907.7µ ± 1%   -2.87% (p=0.002 n=6)
Query/avg_by_(le)(h_1),_range_query_with_1000_steps                    1.832m ± 0%   1.674m ± 1%   -8.64% (p=0.002 n=6)
Query/avg_by_(le)(h_100),_instant_query                                9.498m ± 2%   9.051m ± 1%   -4.71% (p=0.002 n=6)
Query/avg_by_(le)(h_100),_range_query_with_100_steps                   17.91m ± 1%   17.42m ± 1%   -2.76% (p=0.002 n=6)
Query/avg_by_(le)(h_100),_range_query_with_1000_steps                  96.36m ± 4%   89.98m ± 1%   -6.62% (p=0.002 n=6)
Query/avg_by_(le)(h_2000),_instant_query                               172.5m ± 1%   165.1m ± 1%   -4.29% (p=0.002 n=6)
Query/avg_by_(le)(h_2000),_range_query_with_100_steps                  362.3m ± 1%   324.1m ± 1%  -10.53% (p=0.002 n=6)
Query/avg_by_(le)(h_2000),_range_query_with_1000_steps                  2.848 ± 1%    1.717 ± 1%  -39.71% (p=0.002 n=6)
Query/avg_by_(l)(nh_1),_instant_query                                  825.4µ ± 1%   812.5µ ± 1%   -1.56% (p=0.002 n=6)
Query/avg_by_(l)(nh_1),_range_query_with_100_steps                     984.9µ ± 2%   917.1µ ± 1%   -6.89% (p=0.002 n=6)
Query/avg_by_(l)(nh_1),_range_query_with_1000_steps                    2.225m ± 0%   1.857m ± 4%  -16.51% (p=0.002 n=6)
Query/avg_by_(l)(nh_100),_instant_query                                12.14m ± 1%   12.09m ± 0%        ~ (p=0.394 n=6)
Query/avg_by_(l)(nh_100),_range_query_with_100_steps                   27.82m ± 4%   21.35m ± 4%  -23.25% (p=0.002 n=6)
Query/avg_by_(l)(nh_100),_range_query_with_1000_steps                  233.2m ± 2%   123.2m ± 1%  -47.16% (p=0.002 n=6)
Query/avg_by_(l)(nh_2000),_instant_query                               219.8m ± 1%   221.1m ± 1%   +0.57% (p=0.015 n=6)
Query/avg_by_(l)(nh_2000),_range_query_with_100_steps                  695.5m ± 2%   451.3m ± 2%  -35.11% (p=0.002 n=6)
Query/avg_by_(l)(nh_2000),_range_query_with_1000_steps                  5.689 ± 2%    2.953 ± 2%  -48.09% (p=0.002 n=6)
geomean                                                                23.60m        20.31m       -13.94%

                                                            │      Prometheus      │                Mimir                 │
                                                            │         B/op         │     B/op       vs base               │
Query/avg(a_1),_instant_query                                         23.71Ki ± 0%    22.53Ki ± 0%   -4.97% (p=0.002 n=6)
Query/avg(a_1),_range_query_with_100_steps                            24.79Ki ± 0%    23.13Ki ± 0%   -6.72% (p=0.002 n=6)
Query/avg(a_1),_range_query_with_1000_steps                           29.32Ki ± 0%    27.61Ki ± 0%   -5.85% (p=0.002 n=6)
Query/avg(a_100),_instant_query                                       174.3Ki ± 0%    154.1Ki ± 0%  -11.61% (p=0.002 n=6)
Query/avg(a_100),_range_query_with_100_steps                          230.4Ki ± 0%    210.5Ki ± 0%   -8.62% (p=0.002 n=6)
Query/avg(a_100),_range_query_with_1000_steps                         653.7Ki ± 1%    638.7Ki ± 0%   -2.29% (p=0.002 n=6)
Query/avg(a_2000),_instant_query                                      3.016Mi ± 0%    2.618Mi ± 0%  -13.20% (p=0.002 n=6)
Query/avg(a_2000),_range_query_with_100_steps                         4.125Mi ± 2%    3.719Mi ± 0%   -9.85% (p=0.002 n=6)
Query/avg(a_2000),_range_query_with_1000_steps                        12.15Mi ± 1%    12.04Mi ± 1%   -0.92% (p=0.002 n=6)
Query/avg_without_(l)(h_1),_instant_query                             33.10Ki ± 0%    31.61Ki ± 0%   -4.48% (p=0.002 n=6)
Query/avg_without_(l)(h_1),_range_query_with_100_steps                36.90Ki ± 0%    35.16Ki ± 0%   -4.72% (p=0.002 n=6)
Query/avg_without_(l)(h_1),_range_query_with_1000_steps               63.30Ki ± 1%    61.67Ki ± 0%   -2.58% (p=0.002 n=6)
Query/avg_without_(l)(h_100),_instant_query                          1045.0Ki ± 0%    904.4Ki ± 0%  -13.46% (p=0.002 n=6)
Query/avg_without_(l)(h_100),_range_query_with_100_steps              1.355Mi ± 0%    1.220Mi ± 0%   -9.91% (p=0.002 n=6)
Query/avg_without_(l)(h_100),_range_query_with_1000_steps             3.773Mi ± 1%    3.681Mi ± 0%   -2.42% (p=0.002 n=6)
Query/avg_without_(l)(h_2000),_instant_query                          21.39Mi ± 0%    18.67Mi ± 0%  -12.71% (p=0.002 n=6)
Query/avg_without_(l)(h_2000),_range_query_with_100_steps             27.69Mi ± 0%    25.33Mi ± 0%   -8.53% (p=0.002 n=6)
Query/avg_without_(l)(h_2000),_range_query_with_1000_steps           264.99Mi ± 0%    75.29Mi ± 0%  -71.59% (p=0.002 n=6)
Query/avg_without_(le)(h_1),_instant_query                            31.33Ki ± 0%    30.56Ki ± 0%   -2.45% (p=0.002 n=6)
Query/avg_without_(le)(h_1),_range_query_with_100_steps               35.35Ki ± 0%    34.11Ki ± 0%   -3.53% (p=0.002 n=6)
Query/avg_without_(le)(h_1),_range_query_with_1000_steps              61.60Ki ± 0%    60.68Ki ± 1%   -1.50% (p=0.002 n=6)
Query/avg_without_(le)(h_100),_instant_query                         1085.8Ki ± 0%    941.1Ki ± 0%  -13.32% (p=0.002 n=6)
Query/avg_without_(le)(h_100),_range_query_with_100_steps             1.390Mi ± 0%    1.260Mi ± 0%   -9.32% (p=0.002 n=6)
Query/avg_without_(le)(h_100),_range_query_with_1000_steps            3.819Mi ± 1%    3.717Mi ± 0%   -2.67% (p=0.002 n=6)
Query/avg_without_(le)(h_2000),_instant_query                         22.33Mi ± 0%    19.71Mi ± 1%  -11.74% (p=0.002 n=6)
Query/avg_without_(le)(h_2000),_range_query_with_100_steps            28.55Mi ± 0%    28.04Mi ± 2%   -1.78% (p=0.002 n=6)
Query/avg_without_(le)(h_2000),_range_query_with_1000_steps           297.0Mi ± 0%    107.7Mi ± 0%  -63.74% (p=0.002 n=6)
Query/avg(nh_1),_instant_query                                        40.20Ki ± 0%    38.80Ki ± 0%   -3.49% (p=0.002 n=6)
Query/avg(nh_1),_range_query_with_100_steps                          119.15Ki ± 0%    92.36Ki ± 0%  -22.48% (p=0.002 n=6)
Query/avg(nh_1),_range_query_with_1000_steps                          800.1Ki ± 0%    550.4Ki ± 0%  -31.20% (p=0.002 n=6)
Query/avg(nh_100),_instant_query                                      1.102Mi ± 0%    1.037Mi ± 0%   -5.91% (p=0.002 n=6)
Query/avg(nh_100),_range_query_with_100_steps                        10.873Mi ± 0%    5.975Mi ± 0%  -45.05% (p=0.002 n=6)
Query/avg(nh_100),_range_query_with_1000_steps                        99.12Mi ± 0%    50.28Mi ± 0%  -49.27% (p=0.002 n=6)
Query/avg(nh_2000),_instant_query                                     21.46Mi ± 1%    20.11Mi ± 0%   -6.30% (p=0.002 n=6)
Query/avg(nh_2000),_range_query_with_100_steps                        218.3Mi ± 0%    118.8Mi ± 0%  -45.58% (p=0.002 n=6)
Query/avg(nh_2000),_range_query_with_1000_steps                      2010.0Mi ± 0%   1002.5Mi ± 0%  -50.12% (p=0.002 n=6)
Query/avg_by_(l)(h_1),_instant_query                                  31.28Ki ± 0%    30.56Ki ± 0%   -2.30% (p=0.002 n=6)
Query/avg_by_(l)(h_1),_range_query_with_100_steps                     35.34Ki ± 0%    34.06Ki ± 0%   -3.63% (p=0.002 n=6)
Query/avg_by_(l)(h_1),_range_query_with_1000_steps                    61.62Ki ± 0%    60.86Ki ± 1%   -1.23% (p=0.002 n=6)
Query/avg_by_(l)(h_100),_instant_query                               1086.1Ki ± 0%    941.5Ki ± 0%  -13.31% (p=0.002 n=6)
Query/avg_by_(l)(h_100),_range_query_with_100_steps                   1.392Mi ± 0%    1.260Mi ± 1%   -9.44% (p=0.002 n=6)
Query/avg_by_(l)(h_100),_range_query_with_1000_steps                  3.828Mi ± 1%    3.717Mi ± 0%   -2.90% (p=0.002 n=6)
Query/avg_by_(l)(h_2000),_instant_query                               22.37Mi ± 0%    19.66Mi ± 0%  -12.09% (p=0.002 n=6)
Query/avg_by_(l)(h_2000),_range_query_with_100_steps                  28.51Mi ± 0%    28.05Mi ± 2%   -1.62% (p=0.002 n=6)
Query/avg_by_(l)(h_2000),_range_query_with_1000_steps                 296.9Mi ± 0%    107.5Mi ± 0%  -63.79% (p=0.002 n=6)
Query/avg_by_(le)(h_1),_instant_query                                 33.07Ki ± 0%    31.60Ki ± 0%   -4.44% (p=0.002 n=6)
Query/avg_by_(le)(h_1),_range_query_with_100_steps                    36.88Ki ± 0%    35.14Ki ± 0%   -4.73% (p=0.002 n=6)
Query/avg_by_(le)(h_1),_range_query_with_1000_steps                   63.27Ki ± 0%    61.53Ki ± 1%   -2.76% (p=0.002 n=6)
Query/avg_by_(le)(h_100),_instant_query                              1045.5Ki ± 0%    902.9Ki ± 0%  -13.63% (p=0.002 n=6)
Query/avg_by_(le)(h_100),_range_query_with_100_steps                  1.356Mi ± 0%    1.220Mi ± 0%  -10.03% (p=0.002 n=6)
Query/avg_by_(le)(h_100),_range_query_with_1000_steps                 3.773Mi ± 0%    3.677Mi ± 1%   -2.55% (p=0.002 n=6)
Query/avg_by_(le)(h_2000),_instant_query                              21.37Mi ± 0%    18.70Mi ± 0%  -12.50% (p=0.002 n=6)
Query/avg_by_(le)(h_2000),_range_query_with_100_steps                 27.70Mi ± 0%    25.33Mi ± 1%   -8.58% (p=0.002 n=6)
Query/avg_by_(le)(h_2000),_range_query_with_1000_steps               264.88Mi ± 0%    75.33Mi ± 0%  -71.56% (p=0.002 n=6)
Query/avg_by_(l)(nh_1),_instant_query                                 40.31Ki ± 0%    40.17Ki ± 0%   -0.35% (p=0.002 n=6)
Query/avg_by_(l)(nh_1),_range_query_with_100_steps                   119.21Ki ± 0%    93.88Ki ± 0%  -21.25% (p=0.002 n=6)
Query/avg_by_(l)(nh_1),_range_query_with_1000_steps                   800.0Ki ± 0%    552.2Ki ± 0%  -30.97% (p=0.002 n=6)
Query/avg_by_(l)(nh_100),_instant_query                               1.121Mi ± 0%    1.076Mi ± 0%   -4.00% (p=0.002 n=6)
Query/avg_by_(l)(nh_100),_range_query_with_100_steps                  8.464Mi ± 0%    6.017Mi ± 0%  -28.91% (p=0.002 n=6)
Query/avg_by_(l)(nh_100),_range_query_with_1000_steps                 74.98Mi ± 0%    50.50Mi ± 0%  -32.64% (p=0.002 n=6)
Query/avg_by_(l)(nh_2000),_instant_query                              21.88Mi ± 0%    21.07Mi ± 1%   -3.72% (p=0.002 n=6)
Query/avg_by_(l)(nh_2000),_range_query_with_100_steps                 170.6Mi ± 0%    121.2Mi ± 0%  -28.95% (p=0.002 n=6)
Query/avg_by_(l)(nh_2000),_range_query_with_1000_steps                1.517Gi ± 0%    1.014Gi ± 0%  -33.16% (p=0.002 n=6)
geomean                                                               1.913Mi         1.557Mi       -18.62%

                                                            │      Prometheus      │               Mimir                │
                                                            │      allocs/op       │  allocs/op   vs base               │
Query/avg(a_1),_instant_query                                           418.0 ± 0%    390.0 ± 0%   -6.70% (p=0.002 n=6)
Query/avg(a_1),_range_query_with_100_steps                              428.0 ± 0%    395.0 ± 0%   -7.71% (p=0.002 n=6)
Query/avg(a_1),_range_query_with_1000_steps                             454.0 ± 0%    421.0 ± 0%   -7.27% (p=0.002 n=6)
Query/avg(a_100),_instant_query                                        2.330k ± 0%   2.296k ± 0%   -1.46% (p=0.002 n=6)
Query/avg(a_100),_range_query_with_100_steps                           2.743k ± 0%   2.704k ± 0%   -1.44% (p=0.002 n=6)
Query/avg(a_100),_range_query_with_1000_steps                          5.356k ± 0%   5.320k ± 0%   -0.67% (p=0.002 n=6)
Query/avg(a_2000),_instant_query                                       38.92k ± 0%   38.87k ± 0%   -0.12% (p=0.002 n=6)
Query/avg(a_2000),_range_query_with_100_steps                          46.96k ± 0%   46.91k ± 0%   -0.09% (p=0.002 n=6)
Query/avg(a_2000),_range_query_with_1000_steps                         99.00k ± 0%   99.02k ± 0%        ~ (p=0.193 n=6)
Query/avg_without_(l)(h_1),_instant_query                               539.0 ± 0%    521.0 ± 0%   -3.34% (p=0.002 n=6)
Query/avg_without_(l)(h_1),_range_query_with_100_steps                  569.0 ± 0%    546.0 ± 0%   -4.04% (p=0.002 n=6)
Query/avg_without_(l)(h_1),_range_query_with_1000_steps                 726.0 ± 0%    703.0 ± 0%   -3.17% (p=0.002 n=6)
Query/avg_without_(l)(h_100),_instant_query                            12.59k ± 0%   12.56k ± 0%   -0.20% (p=0.002 n=6)
Query/avg_without_(l)(h_100),_range_query_with_100_steps               15.01k ± 0%   14.98k ± 0%   -0.17% (p=0.002 n=6)
Query/avg_without_(l)(h_100),_range_query_with_1000_steps              30.64k ± 0%   30.64k ± 0%        ~ (p=0.420 n=6)
Query/avg_without_(l)(h_2000),_instant_query                           262.4k ± 1%   243.4k ± 0%   -7.25% (p=0.002 n=6)
Query/avg_without_(l)(h_2000),_range_query_with_100_steps              291.4k ± 0%   291.5k ± 0%   +0.03% (p=0.022 n=6)
Query/avg_without_(l)(h_2000),_range_query_with_1000_steps             629.5k ± 0%   605.7k ± 0%   -3.78% (p=0.002 n=6)
Query/avg_without_(le)(h_1),_instant_query                              531.0 ± 0%    504.0 ± 0%   -5.08% (p=0.002 n=6)
Query/avg_without_(le)(h_1),_range_query_with_100_steps                 561.0 ± 0%    530.0 ± 0%   -5.53% (p=0.002 n=6)
Query/avg_without_(le)(h_1),_range_query_with_1000_steps                718.0 ± 0%    686.0 ± 0%   -4.46% (p=0.002 n=6)
Query/avg_without_(le)(h_100),_instant_query                           12.70k ± 0%   12.85k ± 0%   +1.21% (p=0.002 n=6)
Query/avg_without_(le)(h_100),_range_query_with_100_steps              15.12k ± 0%   15.28k ± 0%   +1.05% (p=0.002 n=6)
Query/avg_without_(le)(h_100),_range_query_with_1000_steps             30.76k ± 0%   30.94k ± 0%   +0.60% (p=0.002 n=6)
Query/avg_without_(le)(h_2000),_instant_query                          259.9k ± 1%   250.7k ± 0%   -3.53% (p=0.002 n=6)
Query/avg_without_(le)(h_2000),_range_query_with_100_steps             293.5k ± 0%   301.1k ± 0%   +2.59% (p=0.002 n=6)
Query/avg_without_(le)(h_2000),_range_query_with_1000_steps            635.6k ± 0%   621.5k ± 0%   -2.23% (p=0.002 n=6)
Query/avg(nh_1),_instant_query                                          687.0 ± 0%    656.0 ± 0%   -4.51% (p=0.002 n=6)
Query/avg(nh_1),_range_query_with_100_steps                            2.256k ± 0%   1.723k ± 0%  -23.63% (p=0.002 n=6)
Query/avg(nh_1),_range_query_with_1000_steps                           15.78k ± 0%   10.74k ± 0%  -31.90% (p=0.002 n=6)
…